base transformed to item, movable and location parts
