+++ location.cpp
@@ -1,4 +1,4 @@
-#include "base.h"
+#include "location.h"
 
 
 //---coord---
@@ -19,24 +19,4 @@ void Located::SetCoord(coord coordinates) {
 
 coord Located::GetCoord() const {
     return coordinates;
-}
-
-
-//---ITEM---
-Item::Item(str ID, str name, short level) {
-    this->ID = ID;
-    this->name = name;
-    this->level = level;
-}
-
-str Item::GetId() const {
-    return ID;
-}
-
-str Item::GetName() const {
-    return name;
-}
-
-short Item::GetLevel() const {
-    return level;
 }
